+++ apps/atl_benchmark/atl/blurim.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void blurim(u_int8_t* v,int M,int N,u_int8_t*output);
+void blurim(float* v,int M,int N,float*output);
