+++ apps/atl_benchmark/atl/blurtiles.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void blurtiles(u_int8_t* v,int M,int N,u_int8_t*output);
+void blurtiles(float* v,int M,int N,float*output);
