+++ apps/atl_benchmark/atl/blurtwo.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void blurtwo(u_int8_t* v,int M,int N,u_int8_t*output);
+void blurtwo(float* v,int M,int N,float*output);
