+++ apps/atl_benchmark/atl/im2col.h
@@ -1,3 +1,3 @@
 #include <stdlib.h>
 
-void im2col(u_int8_t* x,u_int8_t* w,int RR,int W,int K,int B,int A,u_int8_t*output);
+void im2col(float* x,float* w,int RR,int W,int K,int B,int A,float*output);
